refactor(pagamento): extrair componentes da página
